Remove deprecated xref versioning schema
In order to prevent the xref pipeline from depending on the
VersioningService repository for a single file, in early August 2018
the schema got copied into ensembl-production and the copy from
VersioningService got deprecated. See ENSCORESW-2728 for details.

Now that the ensembl-production copy has made it into its first
release branch, it is time to get rid of the deprecated one.